+++ src/idc/imgaug/filter/_sub_images.py
@@ -17,7 +17,7 @@ class SubImages(BatchFilter):
 
     def __init__(self, regions: List[str] = None, region_sorting: str = REGION_SORTING_NONE,
                  num_rows: int = None, num_cols: int = None, row_height: int = None, col_width: int = None,
-                 overlap_right: int = None, overlap_bottom: int = None,
+                 overlap_right: int = None, overlap_bottom: int = None, partial_sub_images: bool = False,
                  include_partial: bool = False, suppress_empty: bool = False, suffix: str = DEFAULT_SUFFIX,
                  pad_width: int = None, pad_height: int = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
@@ -40,6 +40,8 @@ class SubImages(BatchFilter):
         :type overlap_right: int
         :param overlap_bottom: the overlap in pixels on the bottom, if no regions defined
         :type overlap_bottom: int
+        :param partial_sub_images: whether to use sub-images that don't have required col_width/row_height
+        :type partial_sub_images: bool
         :param include_partial: whether to include only annotations that fit fully into a region or also partial ones
         :type include_partial: bool
         :param suppress_empty: suppresses sub-images that have no annotations (object detection)
@@ -62,6 +64,7 @@ class SubImages(BatchFilter):
         self.num_cols = num_cols
         self.row_height = row_height
         self.col_width = col_width
+        self.partial_sub_images = partial_sub_images
         self.overlap_right = overlap_right
         self.overlap_bottom = overlap_bottom
         self.include_partial = include_partial
@@ -122,6 +125,7 @@ class SubImages(BatchFilter):
         parser.add_argument("--num_cols", type=int, help="The number of columns, if no regions defined.", default=None, required=False)
         parser.add_argument("--row_height", type=int, help="The height of rows.", default=None, required=False)
         parser.add_argument("--col_width", type=int, help="The width of columns.", default=None, required=False)
+        parser.add_argument("--partial_sub_images", action="store_true", help="Whether to use sub-images that don't have the required --col_width/--row_height", required=False)
         parser.add_argument("--overlap_right", type=int, help="The overlap between two images (on the right of the left-most image), if no regions defined, gets added to the sub-image width.", default=0, required=False)
         parser.add_argument("--overlap_bottom", type=int, help="The overlap between two images (on the bottom of the top-most image), if no regions defined, gets added to the sub-image width.", default=0, required=False)
         parser.add_argument("-s", "--region_sorting", choices=REGION_SORTING, default=REGION_SORTING_NONE, help="How to sort the supplied region definitions", required=False)
@@ -146,6 +150,7 @@ class SubImages(BatchFilter):
         self.num_cols = ns.num_cols
         self.row_height = ns.row_height
         self.col_width = ns.col_width
+        self.partial_sub_images = ns.partial_sub_images
         self.overlap_right = ns.overlap_right
         self.overlap_bottom = ns.overlap_bottom
         self.include_partial = ns.include_partial
@@ -172,6 +177,8 @@ class SubImages(BatchFilter):
             raise Exception("Neither regions nor #rows/cols nor row height/col width specified!")
         if self.region_sorting is None:
             self.region_sorting = REGION_SORTING_NONE
+        if self.partial_sub_images is None:
+            self.partial_sub_images = False
         if self.include_partial is None:
             self.include_partial = False
         if self.suppress_empty is None:
@@ -211,6 +218,7 @@ class SubImages(BatchFilter):
             else:
                 regions = generate_regions(item.image_width, item.image_height,
                                            row_height=self.row_height, col_width=self.col_width,
+                                           partial=self.partial_sub_images,
                                            overlap_right=self.overlap_right, overlap_bottom=self.overlap_bottom,
                                            logger=self.logger())
                 regions_str = regions_to_string(regions, logger=self.logger())